+++ WebViewBinding/src/hb_webview.cpp
@@ -1,4 +1,4 @@
-//Copyright (c) 2023 Eric Lendvai, Federal Way, WA, USA, MIT License
+//Copyright (c) 2024 Eric Lendvai, Federal Way, WA, USA, MIT License
 
 #define LIBRARYVERSION 128
 #define MAXNUMBEROFWINDOWS 100
@@ -22,6 +22,8 @@
 //   unsigned int count;
 // } context_t;
 
+void wv_PingWebView(const char * par_x ,const char *par_y ,void * );
+void wv_GetWebViewBindingVersion(const char * par_x ,const char *par_y ,void * );
 void wv_CloseWebViewWindow(const char * par_x ,const char *par_y ,void * );
 void wv_CallHB(const char * par_x ,const char *par_y ,void * );
 
@@ -92,9 +94,9 @@ int64_t GetWindowNumber()   // To find a iWindowNumber that is unused. Had to cr
     int64_t iPossibleWindowNumber;
     int64_t iWindowNumber = -1;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWindowNumber\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWindowNumber\0") ;
+//  #endif
 
     for (iPossibleWindowNumber = 1; iPossibleWindowNumber <= MAXNUMBEROFWINDOWS; iPossibleWindowNumber++)
     {
@@ -114,9 +116,9 @@ int64_t CreateWebViewWindow(int64_t par_iWindowNumber,int par_iAllowDeveloperToo
     webview_t NewWebViewHandle;
     int64_t iWindowNumber;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in CreateWebViewWindow\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in CreateWebViewWindow\0") ;
+//  #endif
 
     if (par_iWindowNumber <= MAXNUMBEROFWINDOWS) {
 
@@ -136,9 +138,9 @@ int64_t CreateWebViewWindow(int64_t par_iWindowNumber,int par_iAllowDeveloperToo
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t SetWindowTitle(int64_t par_iWindowNumber,char * par_cTite)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in SetWindowTitle\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in SetWindowTitle\0") ;
+//  #endif
     
     if (par_iWindowNumber <= MAXNUMBEROFWINDOWS) {
         webview_set_title(WebViewHandles[par_iWindowNumber],par_cTite);
@@ -169,12 +171,19 @@ int64_t SetWindowPositionAndSize(int64_t par_iWindowNumber,int64_t par_iTop,int6
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t Navigate(int64_t par_iWindowNumber,char * par_cURL)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in Navigate\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in Navigate\0") ;
+//  #endif
     if (par_iWindowNumber <= MAXNUMBEROFWINDOWS) {
         
         TerminateMethod[par_iWindowNumber] = 0;  //Needed in case Made more than one Run for a New Window
+
+        // Still set in case a WebView aware web app, like DataWharf running inside of WharNet
+        webview_bind(WebViewHandles[par_iWindowNumber],"PingWebView",wv_PingWebView,WebViewHandles[par_iWindowNumber]);
+        // webview_bind(WebViewHandles[par_iWindowNumber],"GetWebViewBindingVersion",wv_GetWebViewBindingVersion,WebViewHandles[par_iWindowNumber]);
+        webview_bind(WebViewHandles[par_iWindowNumber],"CloseWebViewWindow",wv_CloseWebViewWindow,WebViewHandles[par_iWindowNumber]);
+        WindowsNumbers[par_iWindowNumber] = (int) par_iWindowNumber;  // The webview_bind function needed an address where a value is stored.
+        webview_bind(WebViewHandles[par_iWindowNumber],"CallHB"            ,wv_CallHB            ,WindowsNumbers+par_iWindowNumber);   // Will be telling to send back the windows par_iWindowNumber, by using and array of handles WindowsNumbers
 
         webview_navigate(WebViewHandles[par_iWindowNumber], par_cURL);
     }
@@ -184,9 +193,9 @@ int64_t Navigate(int64_t par_iWindowNumber,char * par_cURL)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t Run(int64_t par_iWindowNumber)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in Run\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in Run\0") ;
+//  #endif
     if (par_iWindowNumber <= MAXNUMBEROFWINDOWS) {
         webview_run(WebViewHandles[par_iWindowNumber]);
         return TerminateMethod[par_iWindowNumber];
@@ -197,9 +206,9 @@ int64_t Run(int64_t par_iWindowNumber)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t Destroy(int64_t par_iWindowNumber)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in Destroy\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in Destroy\0") ;
+//  #endif
     
     if (par_iWindowNumber <= MAXNUMBEROFWINDOWS) {
         webview_destroy(WebViewHandles[par_iWindowNumber]);
@@ -212,9 +221,9 @@ int64_t Destroy(int64_t par_iWindowNumber)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t Terminate(int64_t par_iWindowNumber)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in Terminate\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in Terminate\0") ;
+//  #endif
     webview_terminate(WebViewHandles[par_iWindowNumber]);
     TerminateMethod[par_iWindowNumber] = 1;
     return 1;
@@ -226,24 +235,59 @@ int64_t SetCallBackFunctionToJSToHarbour(void (*par_ptr)(const char * , const ch
     return 1;
 }
 //---------------------------------------------------------------------------------------------------------------------------------------
+void wv_PingWebView(const char * par_szNumRequests,const char * par_y,void * par_t){
+#if defined(_WIN32)
+    char textBuffer[256];
+    webview_t wv = *( webview_t *)par_t;
+
+    sprintf(textBuffer,"[Harbour] in wv_PingWebView params:[%s],[%s],[%p]\0",par_szNumRequests,par_y,par_t);
+    OutputDebugStringA(textBuffer);
+#endif
+// No return value since the purpose to see if we don't get an error.
+}
+//---------------------------------------------------------------------------------------------------------------------------------------
+//I tried many ways to make this work, but for an unknown reason calling webview_return directly from the received entry point never works. Maybe this is a threading issue.
+//left the code here in case in the future will find a solution.
+void wv_GetWebViewBindingVersion(const char * par_szNumRequests,const char * par_y,void * par_t){
+
+//     #if defined(_WIN32)
+//         char textBuffer[256];
+//     #endif
+//         webview_t wv = *( webview_t *)par_t;
+//     #if defined(_WIN32)
+//         sprintf(textBuffer,"[Harbour] in wv_GetWebViewBindingVersion params:[%s],[%s],[%p]\0",par_szNumRequests,par_y,par_t);
+//         OutputDebugStringA(textBuffer);
+//     #endif
+//     // webview_terminate(wv);
+//     // webview_return(wv, par_szNumRequests, 0,"{\"Version\":\"007\"}\0");
+//     // webview_return(wv, par_szNumRequests, 0,"{\"version\":\"salut\"}\0");
+//     
+//     // sprintf(textBuffer,"{\"version\":\"HelloWorld003\"}\0");
+//     //webview_return(wv, par_szNumRequests, 0,textBuffer);
+//     
+//     //webview_return(wv, par_szNumRequests, 0,"{\"version\":\"HelloWorld003\"}\0");
+//     webview_return(wv, par_szNumRequests, 0,"{\"version\":\"HelloWorld003\"}");
+
+}
+//---------------------------------------------------------------------------------------------------------------------------------------
 void wv_CloseWebViewWindow(const char * par_x,const char * par_y,void * par_t){
 #if defined(_WIN32)
     char textBuffer[256];
 #endif
     webview_t wv = *( webview_t *)par_t;
-#if defined(_WIN32)
-    sprintf(textBuffer,"[Harbour] in wv_CloseWebViewWindow params:[%s],[%s],[%p]\0",par_x,par_y,par_t);
-    OutputDebugStringA(textBuffer);
-#endif
+//  #if defined(_WIN32)
+//      sprintf(textBuffer,"[Harbour] in wv_CloseWebViewWindow params:[%s],[%s],[%p]\0",par_x,par_y,par_t);
+//      OutputDebugStringA(textBuffer);
+//  #endif
     TerminateMethod[FindWebViewHandleNumber(par_t)] = 1;
     webview_terminate(wv);
 }
 //---------------------------------------------------------------------------------------------------------------------------------------
 void wv_CallHB( const char * par_szNumRequests, const char * par_szJson, void * par_p )
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in wv_CallHB\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in wv_CallHB\0") ;
+//  #endif
     
     (*function_JSToHarbour_pointer)(par_szNumRequests,par_szJson,par_p);
 }
@@ -252,9 +296,9 @@ int64_t FindWebViewHandleNumber(webview_t par_WebViewHandle)
 {
     int64_t iWindowNumber = 0;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in FindWebViewHandleNumber\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in FindWebViewHandleNumber\0") ;
+//  #endif
 
     //Find an Unused WebViewHandle
     for (iWindowNumber = 1; iWindowNumber <= MAXNUMBEROFWINDOWS; iWindowNumber++)
@@ -270,9 +314,9 @@ int64_t FindWebViewHandleNumber(webview_t par_WebViewHandle)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t RunJS(int64_t par_iWindowNumber,char * par_cJS)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in RunJS\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in RunJS\0") ;
+//  #endif
     
     webview_eval(WebViewHandles[par_iWindowNumber], par_cJS);
 
@@ -281,13 +325,14 @@ int64_t RunJS(int64_t par_iWindowNumber,char * par_cJS)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t SetHTML(int64_t par_iWindowNumber,char * par_cHTML)
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in SetHtml\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in SetHtml\0") ;
+//  #endif
     
     webview_set_html(WebViewHandles[par_iWindowNumber], par_cHTML);
 
-// bind should be set earlier ? _M_
+    webview_bind(WebViewHandles[par_iWindowNumber],"PingWebView",wv_PingWebView,WebViewHandles[par_iWindowNumber]);
+    // webview_bind(WebViewHandles[par_iWindowNumber],"GetWebViewBindingVersion",wv_GetWebViewBindingVersion,WebViewHandles[par_iWindowNumber]);
     webview_bind(WebViewHandles[par_iWindowNumber],"CloseWebViewWindow",wv_CloseWebViewWindow,WebViewHandles[par_iWindowNumber]);
     WindowsNumbers[par_iWindowNumber] = (int) par_iWindowNumber;  // The webview_bind function needed an address where a value is stored.
     webview_bind(WebViewHandles[par_iWindowNumber],"CallHB"            ,wv_CallHB            ,WindowsNumbers+par_iWindowNumber);   // Will be telling to send back the windows par_iWindowNumber, by using and array of handles WindowsNumbers
@@ -297,9 +342,9 @@ int64_t SetHTML(int64_t par_iWindowNumber,char * par_cHTML)
 //---------------------------------------------------------------------------------------------------------------------------------------
 int64_t InjectJavaScript(int64_t par_iWindowNumber,char * par_cJS)  // Was not able to confirm if this works yet.
 {
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in InjectJavaScript\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in InjectJavaScript\0") ;
+//  #endif
     
     webview_init(WebViewHandles[par_iWindowNumber], par_cJS);
 
@@ -310,9 +355,9 @@ int64_t BringWebViewWindowForeground(int64_t par_iWindowNumber)
 {
     void * iWindowHandle;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in BringWebViewWindowForeground\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in BringWebViewWindowForeground\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
     BringWindowToTop( (HWND) iWindowHandle);
@@ -326,9 +371,9 @@ int64_t GetWebViewWindowPositionTop(int64_t par_iWindowNumber)
     void * iWindowHandle;
     RECT Rect;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWebViewWindowPositionTop\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWebViewWindowPositionTop\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
     GetWindowRect((HWND) iWindowHandle, &Rect);
@@ -344,9 +389,9 @@ int64_t GetWebViewWindowPositionLeft(int64_t par_iWindowNumber)
     void * iWindowHandle;
     RECT Rect;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWebViewWindowPositionLeft\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWebViewWindowPositionLeft\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
     GetWindowRect((HWND) iWindowHandle, &Rect);
@@ -362,9 +407,9 @@ int64_t GetWebViewWindowSizeWidth(int64_t par_iWindowNumber)
     void * iWindowHandle;
     RECT Rect;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWebViewWindowSizeWidth\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWebViewWindowSizeWidth\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
     GetWindowRect((HWND) iWindowHandle, &Rect);
@@ -380,9 +425,9 @@ int64_t GetWebViewWindowSizeHeight(int64_t par_iWindowNumber)
     void * iWindowHandle;
     RECT Rect;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWebViewWindowSizeHeight\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWebViewWindowSizeHeight\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
     GetWindowRect((HWND) iWindowHandle, &Rect);
@@ -392,14 +437,28 @@ int64_t GetWebViewWindowSizeHeight(int64_t par_iWindowNumber)
     return iResult;
 }
 //---------------------------------------------------------------------------------------------------------------------------------------
+int64_t DisableWindowCloseOption(int64_t par_iWindowNumber)
+{
+    void * iWindowHandle;
+
+#if defined(_WIN32)
+    // OutputDebugStringA("[Harbour] in DisableWindowCloseOption\0") ;
+
+    iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
+    EnableMenuItem(GetSystemMenu((HWND) iWindowHandle, FALSE), SC_CLOSE, MF_BYCOMMAND | MF_DISABLED | MF_GRAYED);
+
+#endif
+    return 1;
+}
+//---------------------------------------------------------------------------------------------------------------------------------------
 int64_t MoveWebViewWindow(int64_t par_iWindowNumber,int64_t par_iTop,int64_t par_iLeft,int64_t par_iWidth,int64_t par_iHeight)
 {
     int64_t iResult;
     void * iWindowHandle;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in GetWebViewWindowSizeHeight\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in GetWebViewWindowSizeHeight\0") ;
+//  #endif
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
 
@@ -413,9 +472,9 @@ int64_t AddWebViewWindowTaskBarIcon64(int64_t par_iWindowNumber,char * par_cPath
     HICON hImg;
     void * iWindowHandle;
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in AddWebViewWindowTaskBarIcon\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in AddWebViewWindowTaskBarIcon\0") ;
+//  #endif
     hImg = (HICON) LoadImage(NULL, par_cPath, IMAGE_ICON,64,64,LR_LOADFROMFILE);
 
     iWindowHandle = webview_get_window(WebViewHandles[par_iWindowNumber]);
@@ -428,9 +487,9 @@ int64_t AddWebViewWindowTaskBarIcon64(int64_t par_iWindowNumber,char * par_cPath
 int64_t JSToHarbourReturn(int64_t par_iWindowNumber,char * par_cCallCounter,char * par_cJson)
 {
 
-#if defined(_WIN32)
-    OutputDebugStringA("[Harbour] in JSToHarbourReturn\0") ;
-#endif
+//  #if defined(_WIN32)
+//      OutputDebugStringA("[Harbour] in JSToHarbourReturn\0") ;
+//  #endif
 
     webview_return(WebViewHandles[par_iWindowNumber], par_cCallCounter, 0,par_cJson);